refactor: decouple view and controller
