+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/Catalog/QuickBarView.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c6999d2f124612a4aa6d515c1214bad4
+guid: 63ca2ec1bcf6c364fbd37f25eee7e8d2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
