+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/Catalog/SceneObjectCatalog.asmdef
@@ -12,7 +12,8 @@
         "GUID:4720e174f2805c74bb7aa94cc8bb5bf8",
         "GUID:760a1d365aad58044916992b072cf2a6",
         "GUID:36f22d845f1f3174da2d5fb876f4d3c6",
-        "GUID:cc6bfabbfe87b7949aa248893f89ce37"
+        "GUID:cc6bfabbfe87b7949aa248893f89ce37",
+        "GUID:98bf3bf29030cbf4092d89a7cc28b7fb"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
