+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/BuildModeHUD/BuildModeHUDController.cs
@@ -28,16 +28,19 @@ public class BuildModeHUDController : IHUD
                                                   OnEntityLock,
                                                   OnEntityChangeVisibility;
 
+    
     //Note(Adrian): This is used right now for tutorial purposes
     public event Action OnCatalogOpen;
 
     internal BuildModeHUDView view;
     BuildModeEntityListController buildModeEntityListController;
+
     bool areExtraButtonsVisible = false,isControlsVisible = false, isEntityListVisible = false, isSceneLimitInfoVisibile = false,isCatalogOpen = false;
 
     public BuildModeHUDController()
     {
         view = UnityEngine.Object.Instantiate(Resources.Load<GameObject>("BuildModeHUD")).GetComponent<BuildModeHUDView>();
+
         view.name = "_BuildModeHUD";
         view.gameObject.SetActive(false);
