+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/BuildModeHUD/BuildModeHUDView.cs
@@ -10,7 +10,7 @@ public class BuildModeHUDView : MonoBehaviour
     public SceneLimitInfoController sceneLimitInfoController;
     public SceneObjectCatalogController sceneObjectCatalogController;
     public ToolTipController toolTipController;
-    public QuickBarController quickBarController;
+    public QuickBarView quickBarView;
 
     public GameObject firstPersonCanvasGO, godModeCanvasGO, extraBtnsGO, shortCutsGO;
     public Button firstPersonChangeModeBtn,changeModeBtn,extraBtn,controlsBtn,closeControlsBtn,hideUIBtn,entityListBtn,closeEntityListBtn,catalogBtn,closeCatalogBtn;
@@ -31,15 +31,6 @@ public class BuildModeHUDView : MonoBehaviour
     [SerializeField] internal InputAction_Trigger toggleSceneInfoInputAction;
     [SerializeField] internal InputAction_Trigger toggleCatalogInputAction;
 
-    [SerializeField] internal InputAction_Trigger quickBar1InputAction;
-    [SerializeField] internal InputAction_Trigger quickBar2InputAction;
-    [SerializeField] internal InputAction_Trigger quickBar3InputAction;
-    [SerializeField] internal InputAction_Trigger quickBar4InputAction;
-    [SerializeField] internal InputAction_Trigger quickBar5InputAction;
-    [SerializeField] internal InputAction_Trigger quickBar6InputAction;
-    [SerializeField] internal InputAction_Trigger quickBar7InputAction;
-    [SerializeField] internal InputAction_Trigger quickBar8InputAction;
-    [SerializeField] internal InputAction_Trigger quickBar9InputAction;
 
     public event Action OnControlsVisibilityAction, OnChangeUIVisbilityAction, OnTranslateSelectionAction, OnRotateSelectionAction, OnScaleSelectionAction, OnResetSelectedAction, OnDuplicateSelectionAction, OnDeleteSelectionAction;
     public event Action OnChangeModeAction,OnExtraBtnsClick,OnEntityListChangeVisibilityAction,OnSceneLimitInfoControllerChangeVisibilityAction, OnSceneCatalogControllerChangeVisibilityAction;
@@ -64,19 +55,7 @@ public class BuildModeHUDView : MonoBehaviour
         toggleSceneInfoInputAction.OnTriggered += OnSceneLimitInfoControllerChangeVisibilityTriggered;
         toggleCatalogInputAction.OnTriggered += OnSceneCatalogControllerChangeVisibilityTriggered;
 
-        quickBar1InputAction.OnTriggered += OnQuickBar1InputTriggedered;
-        quickBar2InputAction.OnTriggered += OnQuickBar2InputTriggedered;
-        quickBar3InputAction.OnTriggered += OnQuickBar3InputTriggedered;
-        quickBar4InputAction.OnTriggered += OnQuickBar4InputTriggedered;
-        quickBar5InputAction.OnTriggered += OnQuickBar5InputTriggedered;
-        quickBar6InputAction.OnTriggered += OnQuickBar6InputTriggedered;
-        quickBar7InputAction.OnTriggered += OnQuickBar7InputTriggedered;
-        quickBar8InputAction.OnTriggered += OnQuickBar8InputTriggedered;
-        quickBar9InputAction.OnTriggered += OnQuickBar9InputTriggedered;
-
-       
-
-
+    
         entityListBtn.onClick.AddListener(() => OnEntityListChangeVisibilityAction?.Invoke());
         closeEntityListBtn.onClick.AddListener(() => OnEntityListChangeVisibilityAction?.Invoke());
         
@@ -100,8 +79,8 @@ public class BuildModeHUDView : MonoBehaviour
         deleteBtn.onClick.AddListener(() => OnDeleteSelectionAction?.Invoke());
 
         sceneObjectCatalogController.OnSceneObjectSelected += (x) => OnSceneObjectSelected?.Invoke(x);
-        quickBarController.OnResumeInput += () => OnResumeInput?.Invoke();
-        quickBarController.OnStopInput += () => OnStopInput?.Invoke();
+        quickBarView.OnResumeInput += () => OnResumeInput?.Invoke();
+        quickBarView.OnStopInput += () => OnStopInput?.Invoke();
 
         tutorialBtn.onClick.AddListener(() => OnTutorialAction?.Invoke());
         publishBtn.onClick.AddListener(() => OnPublishAction?.Invoke()); 
@@ -123,16 +102,6 @@ public class BuildModeHUDView : MonoBehaviour
         toggleOpenEntityListInputAction.OnTriggered -= OnEntityListActionTriggered;
         toggleSceneInfoInputAction.OnTriggered -= OnSceneLimitInfoControllerChangeVisibilityTriggered;
         toggleCatalogInputAction.OnTriggered -= OnSceneCatalogControllerChangeVisibilityTriggered;
-
-        quickBar1InputAction.OnTriggered -= OnQuickBar1InputTriggedered;
-        quickBar2InputAction.OnTriggered -= OnQuickBar2InputTriggedered;
-        quickBar3InputAction.OnTriggered -= OnQuickBar3InputTriggedered;
-        quickBar4InputAction.OnTriggered -= OnQuickBar4InputTriggedered;
-        quickBar5InputAction.OnTriggered -= OnQuickBar5InputTriggedered;
-        quickBar6InputAction.OnTriggered -= OnQuickBar6InputTriggedered;
-        quickBar7InputAction.OnTriggered -= OnQuickBar7InputTriggedered;
-        quickBar8InputAction.OnTriggered -= OnQuickBar8InputTriggedered;
-        quickBar9InputAction.OnTriggered -= OnQuickBar9InputTriggedered;
     }
 
     public void SetVisibilityOfCatalog(bool isVisible)
@@ -187,49 +156,8 @@ public class BuildModeHUDView : MonoBehaviour
         toolTipController.Desactivate();
     }
 
-
-    void QuickBarInput(int quickBarSlot)
-    {
-        quickBarController.QuickBarObjectSelected(quickBarSlot);
-    }
     #region Triggers
 
-    private void OnQuickBar9InputTriggedered(DCLAction_Trigger action)
-    {
-        QuickBarInput(8);
-    }
-    private void OnQuickBar8InputTriggedered(DCLAction_Trigger action)
-    {
-        QuickBarInput(7);
-    }
-    private void OnQuickBar7InputTriggedered(DCLAction_Trigger action)
-    {
-        QuickBarInput(6);
-    }
-    private void OnQuickBar6InputTriggedered(DCLAction_Trigger action)
-    {
-        QuickBarInput(5);
-    }
-    private void OnQuickBar5InputTriggedered(DCLAction_Trigger action)
-    {
-        QuickBarInput(4);
-    }
-    private void OnQuickBar4InputTriggedered(DCLAction_Trigger action)
-    {
-        QuickBarInput(3);
-    }
-    private void OnQuickBar3InputTriggedered(DCLAction_Trigger action)
-    {
-        QuickBarInput(2);
-    }
-    private void OnQuickBar2InputTriggedered(DCLAction_Trigger action)
-    {
-        QuickBarInput(1);
-    }
-    private void OnQuickBar1InputTriggedered(DCLAction_Trigger action)
-    {
-        QuickBarInput(0);
-    }
     private void OnSceneCatalogControllerChangeVisibilityTriggered(DCLAction_Trigger action)
     {
         OnSceneCatalogControllerChangeVisibilityAction?.Invoke();
